migrate utility functions to utils.py
