+++ streamlit_app.py
@@ -25,46 +25,6 @@ def init_vectorstore():
     return vectorstore
 
 
-# for extracting source dictionary from retriever results
-def extract_sources(docs):
-    sources = []
-    for doc in docs:
-        source = {
-            "subreddit": doc.metadata["subreddit"],
-            "title": doc.metadata["title"],
-            "content": doc.page_content,
-        }
-        sources.append(source)
-    return sources
-
-
-# for generating formatted html to display sources
-def get_sources_html(sources):
-    n_sources = len(sources)
-    html_out = ""
-    for i, source in enumerate(sources):
-        html_out += f"<p><blockquote>{source['content']}</blockquote></p>"
-        html_out += f"Subreddit: {source['subreddit']}, Title: {source['title']}"
-        if i < n_sources - 1:
-            html_out += "<p><hr/></p>"
-    return html_out
-
-
-# reset chat history in session state and memory object
-def clear_chat_history():
-    st.session_state.messages = [
-        {"role": "assistant", "content": "How may I assist you today?"}
-    ]
-    st.session_state.memory = ConversationSummaryBufferMemory(
-        llm=OpenAI(),
-        memory_key="chat_history",
-        input_key="human_input",
-        max_token_limit=100,
-        human_prefix="",
-        ai_prefix="",
-    )
-
-
 # App title
 st.set_page_config(page_title="Reddit Finance Chatbot 🤑💬")
 
@@ -99,12 +59,18 @@ with st.sidebar:
             "temperature", min_value=0.01, max_value=5.0, value=0.7, step=0.01
         )
         top_p = st.slider("top_p", min_value=0.01, max_value=1.0, value=0.9, step=0.01)
+        model_kwargs = {
+            "model_name": selected_model,
+            "temperature": temperature,
+            "top_p": top_p,
+        }
 
     # metadata filters
     with st.expander("Metadata Filter"):
         subreddits_selected = st.multiselect(
             "subreddit(s) included", options=all_subreddits, default=all_subreddits
         )
+        metadata_filter = {"subreddit": {"$in": subreddits_selected}}
 
     # reset history in session state and memory object
     st.button("Clear Chat History", on_click=clear_chat_history)
@@ -144,49 +110,6 @@ for message in st.session_state.messages:
                 st.write(sources_html, unsafe_allow_html=True)
 
 
-# Function for generating openai response
-def generate_openai_response(human_input):
-    template = """You are a chatbot having a conversation with a human. 
-    You are an expert on the finance opinion from the collective reddit community.
-    Use the following pieces of context to answer the question at the end. 
-    If you don't know the answer, just say that you don't know, don't try to make up an answer. 
-    Use five sentences maximum and explain your reasoning. 
-
-    {context}
-
-    {chat_history}
-
-    Question: {human_input}
-
-    Helpful Answer:"""
-    QA_CHAIN_PROMPT = PromptTemplate(
-        input_variables=["context", "human_input", "chat_history"], template=template
-    )
-
-    # stuff chain
-    llm = OpenAI(model_name=selected_model, temperature=temperature, top_p=top_p)
-    qa_chain = load_qa_chain(
-        llm=llm,
-        chain_type="stuff",
-        prompt=QA_CHAIN_PROMPT,
-        verbose=True,
-        memory=st.session_state["memory"],
-    )
-
-    # generate response
-    similar_docs = vectorstore.max_marginal_relevance_search(
-        human_input,
-        k=4,
-        fetch_k=30,
-        lambda_mult=0.5,
-        filter={"subreddit": {"$in": subreddits_selected}},
-    )
-    result = qa_chain({"input_documents": similar_docs, "human_input": human_input})
-
-    # return result
-    return result
-
-
 # User-provided prompt
 if prompt := st.chat_input(disabled=not openai_key):
     st.session_state.messages.append({"role": "user", "content": prompt})
@@ -197,7 +120,9 @@ if prompt := st.chat_input(disabled=not openai_key):
 if st.session_state.messages[-1]["role"] != "assistant":
     with st.chat_message("assistant"):
         with st.spinner("Thinking..."):
-            result = generate_openai_response(prompt)
+            result = generate_openai_response(
+                prompt, vectorstore, model_kwargs, metadata_filter
+            )
             st.write(result["output_text"])
             sources = extract_sources(result["input_documents"])
             with st.expander("Sources"):
